Initial commit of Dashboard API implementation of ruby. End goal is to convert this into a Gem instead of just some files.
File structure is going to be pretty straight forward:
```
├── lib
│   ├── dashboard_api.rb
│   └── organizations.rb
```
`dashboard_api.rb` is the main DashboardAPI class that you will actually instantiate.
`organizations.rb` is a module that represents the "Organizations" section of the Meraki Dashboard API documentation that is located from directly on Dashboard
under Help -> API Docs.

Each section under Help -> API Docs will be split into it's own module for convenience. Eventually tests will be moved to that convention, with a rake task for each section as needeed.

Because Dashboad API requires an API key, as well as unique values for certain API called (org_id, network_id, etc.), environment variables are required to be set. Currently only

* `dashboard_api_key`
* `dashboard_org_id`

need to be set if you want to develop / test the codebase. No ENVs need to be set for normal usage.