Added POST functionality (details below)
dashboard_api:
Now contains support for POST functionality. The actual options that will be passed during the post, such as
the network name and type when creating a new network are part of the 'body' section and automatically gets
serialized to json.

The POST will also watch out for errors, such as a network already existing, and will raise an exception for this
letting you know why it errored.

test_dashboard_api:
3 new tests. 1 to check basic functionality of the make_api_call for GET
1 to check the basic functionality of make_api_call for POST
1 to check the basic error handling of make_api_call for POST

organizations.rb
Since make_api_call now requries an HTTP_METHOD to be passed along side of it, all previous calls needed
to be updated appropriately