Started writing for the Networks section. Basic functionality up to the point of needing templates is available.
With this comes full HTTP METHOD functionality for GET, POST, PUT and DELETE.

Tests have been modified to not check for specific things in the responses, as this would break tests when building on someone elses machines, and their fixtures are not the same. Instead we will trust that if a non-404 response comes back, and it returns as the proper data type, it is the proper data.